Set VERSION and SONAME version properties from CMake
If clsocket is compiled dynamically, this will produce .so files
with the version in their name, e.g. libclsocket.so.1.4.3, to which
libclsocket.so.1 and libclsocket.so is symlinked to. This means that
if anyone wants to use this dynamically, they will nominally be able
to better ensure ABI compatibility.

Since the version was already included in the CMake configuration,
this is a very simple change.

Since dfhack is linking against clsocket statically, this shouldn't
affect it at all, but this change should help anyone who wants to use
clsocket as a dynamic library (or build dfhack against a dynamic
clsocket).